requirements.txt
black==20.8b1
dj-database-url==0.5.0
django>=3.1.8
django-heroku==0.3.1
factory-boy==3.1.0
django-credit-cards==0.4.1
Pillow~=8.1.0
gunicorn==20.1.0
psycopg2